Modify Operator Conditions (#130)
This PR brings in 2 changes:

1. Introduce a new reason for Unknown status:
Currently, we specify BundleDeploymentFailure or InstallationSuccess as reasons
for an unknown status. This means that we are not sure if the install is successfull or
we aren't sure of a failure either. To make it more accurate, we introduce a new reason
"InstallStatusUnknown" to mean that the status cannot be determined at that particular instant.

2. Rename "BundleDeploymentFailed" to "InstallationUnsuccessful"
This prevents leaking of Rukpak API names. BundleDeployment Failed can be one of the reason
for unsuccessful installation.

Signed-off-by: Varsha Prasad Narsing <varshaprasad96@gmail.com>