Fixes kind cluster cleanup in `e2e` make target
Currently we have `kind-cluster-cleanup` as a dependency
in `kind-cluster` target. Since `make` only builds a target
once this means that `e2e` target performs cleanup
before creating a new cluster and skips cleanup at the end
leaving cluster hanging around.

With this change, `make e2e` will be failing on cluster creation,
if a cluster with the same name already exists, but the failure
message will clearly indicate the issue.

Signed-off-by: Mikalai Radchuk <mradchuk@redhat.com>