Remove go-apidiff GHA from merge_group runs (#373)
go-apidiff is meant to be informational and is optional, so it will not block PRs in the merge queue, even when it finds incompatible changes. Therefore it serves no real purpose in the merge queue CI.

There also seems to be an issue with running go-apidiff in the context of a merge_queue where it fails due to invalid commit references.

For these reasons, removing it from the merge queue CI

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>