Fixes reading properties with the same type (#427)
Some properties like `olm.package` only can appear in a bundle once
and we currently handle this correctly.

However properties such as `olm.gvk` and `olm.gvk.required`
can be present multiple times in the list. The current
implementation overrides all the instances with the one which is
the last in the list.

This modifies the code to be able to read in both cases.

Signed-off-by: Mikalai Radchuk <mradchuk@redhat.com>