Update E2E test data
Zero major version is reserved for initial development
according to SemVer spec[^1] and anything may change at any time.
Updates for this version behaves differently:

* When version is `0.1.0` we allow versions which are >= `0.1.0` and < `0.2.0` (patch updates)
* When version is `0.0.1` we allow versions which are >= `0.0.1` and < `0.0.2` (no updates)

To be able to update in the second case (e.g. from `0.0.1` to `0.0.2`)
users will have to set `.spec.upgradeConstraintPolicy` to `Ignore`
once support for this implemented.

We have unit tests for the above behaviour, but E2E test data was not updated.

[^1]: https://semver.org/#spec-item-4

Signed-off-by: Mikalai Radchuk <mradchuk@redhat.com>