Update installed package unit tests
This commit does two things.

First: It splits the `TestMakeInstalledPackageVariables` test into two tests:
* One for when `ForceSemverUpgradeConstraints` feature gate enabled (semver)
* One for when `ForceSemverUpgradeConstraints` feature gate disabled (legacy semantics)

Both tests are now table-style tests.

This is done mainly so that we can maintain test data for different semantics separately (e.g. no need to maintain channel entires for semver tests).

Second: Adds extra coverage for disabled state of the `ForceSemverUpgradeConstraints`
feature gate. Previously we were not covering some cases for this state of the gate.
Example of such case is when `UpgradeConstraintPolicy` field on `Operator` is set to `Ignore`.

Signed-off-by: Mikalai Radchuk <mradchuk@redhat.com>