Use bin-dir override on setup_envtest (#790)
Rather than change XDG_DATA_HOME, since many distributions do not
explicitly define XDG_DATA_HOME but rather fallback to the default,
check if XDG_DATA_HOME is not set and if HOME is either not set or set
to the root filesystem. If so, override the setup_envtest binary
directory. This helps prevent errors on some CI/CD environments.

Signed-off-by: Tayler Geiger <tayler@redhat.com>