[Fix] Fix constant reconciles due to Catalog (#812)
Due to the polling feature in the Catalog, the `lastPolledInterval` in the
Catalog's status gets updated. Because of this there is a reconcile is trigrred
every time the catalog is polled, even though there is no change in the
resolved image ref SHA.

This causes unnecessary continuous reconciles in the CE. To overcome this,
a Predicate is added to accept update events only when there is a change in
Resolved reference.

Signed-off-by: Varsha Prasad Narsing <varshaprasad96@gmail.com>