Fix: installed bundle provider (#916)
The installed bundle provider, which is necessary for safely handling
package upgrades, was too strict. It required the currently installed
bundle to exist in the catalog in order to "find" the installed bundle.

This is problematic in several  situations:
- If a user receives a catalog update that no longer contains their
  installed bundle (perhaps its was pulled from the catalog for security
  or policy reasons)
- If a user is trying to transition to a different catalog that provides
  that package
- If a bundle was directly installed, and the user is trying to have a
  catalog-backed ClusterExtension adopt it.

This change simply returns the name and version of the installed bundle,
and makes some minor changes in the successors functions to handle the
simplified installed bundle metadata.