fix: reduce unnecessary caching and remove predicate from dynamic watches (#963)
Our cache configuration resulted in watching all cluster-scoped
resources for each cluster-scoped resource that shows up in a
ClusterExtension. Now, we only watch:
- All ClusterExtensions
- All ClusterCatalogs
- All objects in the systemNamespace
- All other objects that specifically match the dependentPredicate

This commit also removes the dependent predicate so that we reconcile
for any change to managed resources. This is important to ensure that
future features (like health checking) can account for any change to
managed operator objects.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>