Improve Catalog Selection Testing + Fix small issue (#1174)
We initially assumed that the resolvedBundle index we check doesn't matter when it comes to checking for prior package deprecations in the resolver but the index needs to match the prior result or the bundle names may not match and give a false negative.

Also improved the unit test cases (which caught the above issue).

Signed-off-by: dtfranz <dfranz@redhat.com>