Replace `require` with `assert` in retry funcs
When `require` assertions are not satisfied they fails
a test immediately. This is not desired when used
with retry function such as `EventuallyWithT` since
they are supposed to retry on a failure.

This can cause flakes when, for example, we are asserting
on a resource condition. On the first hit the resource might
not have a condition and `require` will fail `EventuallyWithT`
despite the fact that the code is correct and the condition
gets set eventually.

Signed-off-by: Mikalai Radchuk <mradchuk@redhat.com>