Add testify linter and address fixes
While reviewing https://github.com/operator-framework/operator-controller/pull/1330
we had a discussion about the ability to generally prevent the use of `requires`
within `Eventually` calls to prevent future flakes, and that led us to evaluate
the `testifylint`-er to give us some cautionary guidelines.

It did _not_ resolve the requires-within-eventually issue,
but we may try to tackle that with some custom AST in the future.

Signed-off-by: Jordan Keister <jordan@nimblewidget.com>