use spec's ref instead of canonical ref when copying catalog image (#438)
When used in conjuction with a "tag-only" registries.conf mirror
registry configuration, our containers/image unpacker fails. This is
because we resolve a canonical ref from the tag, and then use the
digest-based reference to actually copy the image, thus thwarting the
tag-only configuration.

This commit ensures that we always copy the image using the reference
provided in the spec.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>