(ci): Add GitHub Action to validate installation on Linux and Mac envs
We are currently facing issues where the installation process is breaking, and we are unable to detect these failures in CI. This action introduces a validation step to ensure that we can at least install the solution, validating the installation script provided in the Getting Started guide.

This is a minimal basic test expected to enhance the developer experience (DX) by reducing user frustration with installation issues.