Change formatting in the "Catalog queries" doc (#1461)
This commit adds definition lists for placeholders such as `<package_name>`.
The goal of adding definition lists for placeholders is to improve their
visibility in the code snippets.

This commit also switches existing definition lists [1] (list of queries)
to unordered list. The current list of queries uses definition lists
from mkdocs-material. However it relies on a tailing space
after a colon to work. E.g. the whole line becomes `: `
and it is prone to errors because:
1. It is easy to overlook a space which is part of markdown.
2. Many editors remove tailing spaces at the end of lines.
3. We are adding more definition lists into this docs
   to explain placeholders such as `<package_name>`.
   With nested definition lists markdown of this document
   going to be very difficult to manage.

[1]: https://squidfunk.github.io/mkdocs-material/reference/lists/#using-definition-lists

Signed-off-by: Mikalai Radchuk <mradchuk@redhat.com>