Add example manifest for ArgoCD v0.2.0
For our examples to work, we need to use a very specific version of our
example operator. The upgrade from 0.2.0 to 0.2.1 functions properly for
our purposes, but we don't want to affect the sample operator in
config/samples so we are adding the example manifest for 0.2.0 directly
to this tutorial.

Signed-off-by: Tayler Geiger <tayler@redhat.com>