📖 Update extension upgrade tutorial doc (#1469)
* Update extension upgrade tutorial doc

With the addition of the CRD Upgrade Safety preflight check, our
existing example of upgrading argocd from 0.5.0 to 0.6.0 no longer
serves as a good example. This commit changes the example to use an
update from 0.2.0 to 0.2.1 which no longer fails the CRD Upgrade check,
and also doesn't violate our restriction on automatic upgrades between
minor versions with a major version of zero.

Signed-off-by: Tayler Geiger <tayler@redhat.com>

* Add example manifest for ArgoCD v0.2.0

For our examples to work, we need to use a very specific version of our
example operator. The upgrade from 0.2.0 to 0.2.1 functions properly for
our purposes, but we don't want to affect the sample operator in
config/samples so we are adding the example manifest for 0.2.0 directly
to this tutorial.

Signed-off-by: Tayler Geiger <tayler@redhat.com>

---------

Signed-off-by: Tayler Geiger <tayler@redhat.com>