improve the verify-crd-compatibility make target (#1499)
for use during local development by making the default updated source
be the local file, picking up changes not yet committed. The previous
git sourcing required changes to be committed to be picked up and is not
a very good developer experience as you should validate the changes
prior to commiting them. Since the git source is necessary for CI,
the GitHub Action is updated to specify the git source for the updated file.

Signed-off-by: everettraven <everettraven@gmail.com>