(catalogd) `serveJSON` lines instead of `http.serverContent` for no-params (#1725)
The metas endpoint was using `serverJSONLines` for serving queries that are
parameterized, which copies content from a reader to the response writer under
the hood. As a result no-parameterized query responses don't have range request
support.

The endpoint was however using `http.ServeContent` for cases when no parameters
were provided, which does have range request support.

This PR switches the `http.ServeContent` out for `serveJSONLines` to make sure
metas endpoint is consistent in it's lack of support for range requests.

Signed-off-by: Anik Bhattacharjee <anbhatta@redhat.com>