Handle interrupted helm releases in applier
Introduces a workaround for 'interrupted' helm releases
which enter into a 'pending' (-install/uninstall/rollback) state.
If that happens, for example because of immediate application
exit with one of those operations being in flight, helm is not
able to resolve it automatically which means we end up in
a permanent reconcile error state.
One of the workarounds for this that has been repeated in the
community is to remove metadata of the pending release,
which is the solution chosen here.

For full context see:
https://github.com/operator-framework/operator-controller/pull/1776
https://github.com/helm/helm/issues/5595