Consolidate configuration (#1790)
This moves all the configuration into the root config directory.
The config/README.md file shows the result.

Updated Makefile, goreleaser, documentation and some of the unit test code.

Kept CRDs in their original locations, but made copies to the new
locations to keep the verify-crd-compatibility target working properly.
Once this is merged, a followup will remove the CRDs from the original
locations and update the verify-crd-compatibility target.

It also tries to make catalogd less of a second-class piece of code.

I compared the resulting operator-controller.yaml with main, and with
the exception of `image` locations, they are identical.

Signed-off-by: Todd Short <tshort@redhat.com>