(e2e) fix namespaces network-policy test searches in (#2034)
The test was previously listing networkpolicies to validate only in
the namespace operator-controller is deployed in.
Upstream, that is valid since the namespace that operator-controller is
deployed in, is the namespace that all other components are deployed in
(and therefore all the network polices we expect to find).

A different distribution of olmv1 (downstream) could decide to deploy the
policies in different namespaces, in which case the test will not look into
all the namespaces it should look into.

This PR fixes the issue by searching in the namespaces both operator-controller
and catalogd is deployed in.