Add standard-e2e manifest
This renames two overlays:
* cert-manager -> standard
* e2e -> standard-e2e

This also adds a new manifest:
* manifests/standard-e2e.yaml

This fixes the issue with the current test-e2e modifying the standard
manifest with e2e parameters, leaving a dirty workspace. This changes
the manifest used by the e2e test to be it's own standard-e2e manifest.

This is part of the feature-gated API functionality.

Signed-off-by: Todd Short <tshort@redhat.com>