Rename tls component to cert-manager and consolidate
This renames the tls component to cert-manager to reflect what it really
is. A kustomization.yaml file is added to cert-manager to make it
behave as a single component. This will make it easier to manage this
component, and allow for easier future replacement.

This is part of the feature-gated API functionality.

Signed-off-by: Todd Short <tshort@redhat.com>