Add kustomization files for each CRD (#2061)
At the root of each CRD directory, there is a kustomization file.
Replicate this into the individual directories, so that they can be
referenced directly by overlays.

This keeps the "default" reference of `config/base/.../crd` to be the
standard CRD for compatibility with existing overlays.

The kustomizeconfig.yaml file is not used, and is deleted.

This is part of the feature-gated API functionality.

Signed-off-by: Todd Short <tshort@redhat.com>