Add feature-set annotation to all manifest resources (#2073)
The feature-set annotation was originally applied by the crd-generator
to the generated CRDs. Now, that annotaion is being added by the overlay,
and is being added to all resources in the manifest.

Each manifest has an annotation value based its name.

The crd-generator is now adding a generated annotation indicating what
the CRD was generated for.

Signed-off-by: Todd Short <tshort@redhat.com>