Move e2e registry resources into config
Currently, the e2e registry resources are located in a script
`testdata/build-test-registry.sh`

The resources in this script can be moved the the e2e config and
manifests that now exist.

The only thing that remains is the job, that kinda-sorta needs
to wait until the registry is up and working.

Signed-off-by: Todd Short <tshort@redhat.com>