Remove use of namespace in kustomize
The use of the namespace parameter is kustomization files is very tricky.
In one particular instance, we have an ordering issue with the cert-manager
CA component. If not ordered correctly in the set of kustomization files,
the CA component namespace will be overwritten by prior namespace
directives.

This eliminates that edge case, and makes the kustomization more robust.

Downstream uses a different overlay, so there's no issue there.

Also, add `-n` option to the install script to allow users to easily
change the namespace that they install OLMv1 into.

Note that the manifests don't change; so this keep everything as-is.

Signed-off-by: Todd Short <tshort@redhat.com>