Define fine-grained owners for the various subcomponents of OLMv1.
The goal is to reduce the possibility of the bystander effect
(see https://en.wikipedia.org/wiki/Bystander_effect), and give
maintainers more accountability and ownership of the areas in
which they are experts.

This will also help contributors more quickly identify those experts
and get the necessary reviews for their work to merge.