Remove kaniko / extension-developer-e2e cleanup (#2116)
Removes use of kaniko, which is now an archived project, in favor of using tools we already have at our disposal. Also attempted to clean up the Makefile and setup script a bit, as well as a fail-early check in the test to make it clearer that some env is required.

Signed-off-by: Daniel Franz <dfranz@redhat.com>