🐛 CRD upgrade safety fixes and ratcheting (#2123)
* move crd upgrade safety testdata into crdupgradesafety package

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>

* fix: bump crdify to fix bugs and regressions

regression fixes:

1. correctly handle processing of properties that are OpenAPI items
2. allow enums to have values added.

bug fix:

crdify's served version validator was updated to actually compare the
old CRD with the new CRD so that any issues identified in the old CRD
do not continue to be reported when performing comparisons between
served versions of the new CRD.

This effectively allows issues in the served version validations to be
acknowledged once when they are introduced, but then those issues are
essentially grandfathered in such that they do not have to be
acknowledged again in the future.

This issue was actually identified in a case where an operator upgrade
was stopped by the CRD upgrade check despite there being no changes
whatsoever to the CRD. The "old" and "new" CRDs contained the exact
same issues, but since crdify was looking exclusively at the "new" CRD,
it found those issues and reported them.

---------

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>