Use control-plane selectors in network-policies and tests for now (#2218)
Until downstream is ready to use the "app.kubernetes.io/name" selector,
continue to use the "control-plane" selector in the tests.

Change the network-policies to use a "control-plane" selector (which is still
on pods because the Deployment selector is immutable).

This includes a revert of "Use old and new pod selectors during kustomize-to-helm transition"
This reverts #2214
This reverts commit 6e22e2b0595176c02df054566fc2b0c1f7fd3591.