Add support for TLS profiles
Use Mozilla's profiles to define TLS profiles for operator-controller
and catalogd. These are configured via command-line options, and can
be customized.

The idea is that downstream, cluster-olm-operator will be able to
glean the appropriate configuration, and provide that to the
components.

There is a semi-automatic method to update the profiles, if that
ever happens (`make update-tls-profiles`).

This adds `gojq` via bingo, which is a golang implementation of jq
for the update-tls-profiles target.

Signed-off-by: Todd Short <tshort@redhat.com>