Support disabling feature-gates (#2271)
The Helm manifests only support _enabling_ feature-gates. To allow
for flexibility in configuring mutually-exclusive feature-gates
and additional parameters that might be needed, add Helm templating
for disabled feature gates and extra arguments to the deployments.

NOTE: This deprecates the old location of feature-gate config, and
moves it to a new location. This is necessary to keep downstream
from breaking during this transition.

Signed-off-by: Todd Short <tshort@redhat.com>