:seedling: Use Patch instead of Update for finalizer operations (#2328)
* Use Patch instead of Update for finalizer operations

Refactor all controllers to use Patch() instead of Update()
when adding or removing finalizers to improve performance, and to avoid
removing non-cached fields erroneously.

This is necesary because we no longer cache the last-applied-configuration
annotation, so when we add/remove the finalizers, we are removing that field
from the metadata. This causes issues with clients when they don't see that
annotation (e.g. apply the same ClusterExtension twice).

Update ClusterCatalog and ClusterExtension controllers to use Patch-based
funalizer management (ClusterExtensionRevision already uses Patch())

Signed-off-by: Todd Short <tshort@redhat.com>

* Revert "Restore last-applied-config annotation in cache (#2338)"

This reverts commit 39cbdbe26dbb38444d6489d9bc1fc6e0777f45ec.

---------

Signed-off-by: Todd Short <tshort@redhat.com>