Add documentation URLs into CRD descriptions
These documentation URLs are templated, so that they can be updated
downstream as needed.

Added **quick-helm** utility to go template a file using helm-like
semantics (i.e. everything is prefixed with `.Values`)

This is used to update the docs/api-reference/olmv1-api-reference.md
file, otherwise, Go templates are left in the file.

Signed-off-by: Todd Short <tshort@redhat.com>