Change the experimental-e2e to use a two-node kind cluster (#2382)
In order to fully support multiple replicas, we need to be able to run
our tests in a multi-node environment, but we don't _always_ want to
run in a multi-node test environment.

Update the experimental-e2e to run a 2-node kind cluster (both control-
plane).

Update the `kind-load` recipe to directly load a docker image via
`load docker-image`, since `load image-archive` has issues with
multi-node clusters.

Signed-off-by: Todd Short <tshort@redhat.com>